--- src/components/header.js
@@ -1,10 +1,0 @@
-import logo from '../planet.png';
-
-const Header = () => (
-  <div className="headerSection">
-    <img src={logo} alt="logo" className="logoImg" />
-    <h1>Space Traveler&apos;s Hub</h1>
-  </div>
-);
-
-export default Header;
