+++ src/components/nav/Header.js
@@ -1,0 +1,10 @@
+import logo from '../../planet.png';
+
+const Header = () => (
+  <div className="logo">
+    <img src={logo} alt="logo" className="logo-img" />
+    <h1>Space Traveler&apos;s Hub</h1>
+  </div>
+);
+
+export default Header;
