+++ src/pages/Profile.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Profile = () => <div>Profile Page</div>;
+
+export default Profile;
